Create vocabulary dto

+++ src/main/java/com/learnenglish/studentsvocabulary/repository/VocabularyRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VocabularyRepository extends JpaRepository<Vocabulary, Integer> {
+public interface VocabularyRepository extends JpaRepository<Vocabulary, Long> {
 }
